fix: refactor use search params
