+++ apps/dashboard/src/types/activity.ts
@@ -1,0 +1,15 @@
+import { ChannelTypeEnum } from '@novu/shared';
+import { IActivityFilters } from '@/api/activity';
+
+export interface IActivityFiltersData {
+  dateRange: string;
+  channels: ChannelTypeEnum[];
+  templates: string[];
+  searchTerm: string;
+}
+
+export interface IActivityUrlState {
+  activityItemId: string | null;
+  filters: IActivityFilters;
+  filterValues: IActivityFiltersData;
+}
